requirements.txt
python-binance
python-telegram-bot
pandas
python-dotenv
